Add float,double type to inflator.

+++ tests/schema/tests/AuthorSchemaProxy.php
@@ -70,8 +70,8 @@ class AuthorSchemaProxy extends RuntimeSchema
   'r' => array( 
       'name' => 'r',
       'attributes' => array( 
-          'type' => 'float',
-          'isa' => 'float',
+          'type' => 'double',
+          'isa' => 'double',
         ),
     ),
 );
